Backend/requirements.txt
absl-py==2.1.0
astunparse==1.6.3
backports.tarfile==1.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
docutils==0.20.1
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==24.3.25
fonttools==4.54.1
gast==0.4.0
google-auth==2.35.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.67.0
h5py==3.11.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jax==0.4.13
Jinja2==3.1.4
joblib==1.4.2
keras==2.12.0
keyring==25.4.1
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplot==0.1.9
matplotlib==3.7.5
mdurl==0.1.2
ml-dtypes==0.2.0
more-itertools==10.5.0
nh3==0.2.18
numpy==1.23.5
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==24.1
pandas==2.0.3
pillow==10.4.0
pip==24.2
pkginfo==1.11.2
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.18.0
pyloco==0.0.139
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
readme_renderer==43.0
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.3
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
setuptools==41.2.0
SimpleWebSocketServer==0.1.2
six==1.16.0
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.12.0
termcolor==2.4.0
threadpoolctl==3.5.0
twine==5.1.1
typing==3.7.4.3
typing_extensions==4.5.0
tzdata==2024.2
urllib3==2.2.3
ushlex==0.99.1
websocket-client==1.8.0
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.14.1
zipp==3.20.2
